backend/requirements.txt
Flask==2.3.2
flask_sqlalchemy==3.0.3
pdfminer.six==20231012
spacy==3.6.1
openai==1.0.0
python-dotenv==1.0.0
scikit-learn==1.4.2
